refactor freeform-line
